+++ BlockzillaPackage/Sources/Onboarding/Handler/Action.swift
@@ -1,0 +1,15 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+public enum Action {
+    case applicationDidLaunch
+    case enterHome
+    case showTrackingProtection
+    case trackerBlocked
+    case showTrash
+    case clearTapped
+    case startBrowsing
+}
