+++ BlockzillaPackage/Sources/Onboarding/Handler/OnboardingEventsHandling.swift
@@ -1,0 +1,11 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+public protocol OnboardingEventsHandling: AnyObject {
+    var route: ToolTipRoute? { get set }
+    var routePublisher: Published<ToolTipRoute?>.Publisher { get }
+    func send(_ action: Action)
+}
