+++ BlockzillaPackage/Sources/Onboarding/Handler/ToolTipRoute.swift
@@ -1,0 +1,17 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+public enum ToolTipRoute: Equatable, Hashable, Codable {
+    case onboarding(OnboardingVersion)
+    case trackingProtection
+    case trackingProtectionShield
+    case trackingProtectionShieldV2
+    case trash
+    case trashV2
+    case searchBar
+    case widget
+    case menu
+}
